Don't create dummy data for each date query test.
The integration tests for using date_query in WP_Query require the creation of
dummy post data, much of which is common between a number of tests. However,
newer tests do not require as much data, so to reduce overhead we create the
data explicitly only when we need it, instead of in the setUp() method.

See #29781.
Built from https://develop.svn.wordpress.org/trunk@29885

git-svn-id: http://core.svn.wordpress.org/trunk@29641 1a063a9b-81f0-0310-95a4-ce76da25c4cd