Customizer: Improve previewing setting changes for `show_on_front`, `page_on_front`, and `page_for_posts`.
When changing the `page_on_front` setting (with `show_on_front`), change the previewed URL to be the home URL so that the effect can be seen. When changing `page_for_posts`, change the previewed URL to be the selected page.

Props valendesigns, westonruter.
Fixes #30677.

Built from https://develop.svn.wordpress.org/trunk@32976

git-svn-id: http://core.svn.wordpress.org/trunk@32947 1a063a9b-81f0-0310-95a4-ce76da25c4cd